Remove MsTimer2 lib and update MyoControl lib
The MsTimer2 lib cannot be called from inside another library (actually
it can, but doing some ugly coding tricks). For this reason, the
sampling function has to be called from the main code. To avoid this,
the MyoControl lib has been modified to call the sampling function every
millisecond without using the MsTimer2 lib. Probably, this will result
in a less precise sampling of the EMG signals, but it might work.