Fix sampling function
The sampling function never gave a value of "true" to the sampleOk variable, rendering the rest of the code useless. The "if" condition is supposed to check if the elapsed time between EMG samples is greater than 1ms to take a new sample. However, the function was measuring the time difference between the start of the current execution of the function and the end of the previous execution, not the previous sample. For this reason, the time difference was never greater thatn 1ms. By assigning the value of the nowTime variable to the prevTime variable inside the condition that checks if the elapsed time after the last EMG sample is greater than 1ms, the sampling function works as expected.